Improved contrast for buttons in light mode as well as refactored the code to a new class.
